Move file path resolution to later in the parsing pipeline and fix rollup to ensure react is excluded
