+++ rollup.config.mjs
@@ -7,10 +7,19 @@ import postcss from "rollup-plugin-postcss";
 import peerDepsExternal from "rollup-plugin-peer-deps-external";
 import preserveDirectives from 'rollup-preserve-directives';
 
+
+const external = [
+  'react',
+  'react-dom',
+  'react-gauge-component',
+  'react-redux',
+  '@reduxjs/toolkit'
+];
+
 const config = [
   {
     input: "src/index.ts",
-
+    external,
     output: [
       {
         dir: "dist",
@@ -58,7 +67,7 @@ const config = [
     input: "dist/index.d.ts",
     output: [{ file: "dist/index.d.ts", format: "esm" }],
     plugins: [dts()],
-    external: [/\.css$/, '@reduxjs/toolkit']
+    external: [/\.css$/, ...external]
   }
 ];
