+++ vite.config.ts
@@ -1,13 +1,15 @@
 import { defineConfig, loadEnv } from 'vite';
 import { nodePolyfills } from "vite-plugin-node-polyfills";
 import react from '@vitejs/plugin-react';
-import eslint from 'vite-plugin-eslint';
 
 // https://vitejs.dev/config/
 export default defineConfig(({ command, mode }) => {
     const env = loadEnv(mode, process.cwd(), '');
     return {
-        plugins: [react(), eslint(), nodePolyfills()],
+        plugins: [react(), nodePolyfills()],
+        resolve: {
+            dedupe: ['react', 'react-dom']
+        },
         test: {
             environment: 'jsdom',
             include: ['**/*.test.ts', '**/*.test.tsx'],
